fix: progressBar 컴포넌트 추가 및 StrengthWeaknessFunnel에서 사용
