+++ src/domains/StrengthWeakness/components/StrengthWeaknessFunnel.tsx
@@ -1,6 +1,7 @@
 import { useFunnel } from '@use-funnel/browser';
 import PageBackground from '@/shared/ui/components/PageBackground';
 import userCardSteps from '../hooks/useCardFunnel';
+import ProgressBar from './ProgressBar';
 import StrengthCardStep from './step/StrengthCardStep';
 import StrengthCardStep2 from './step/StrengthCardStep2';
 import WeaknessStep from './step/WeaknessStep';
@@ -14,21 +15,6 @@ const options = {
   } as const,
   steps: userCardSteps,
 };
-
-function ProgressBar({ step, limit = 4 }: { step: number; limit?: number }) {
-  const getWidth = () => {
-    return `${(step / limit) * 100}%`;
-  };
-
-  return (
-    <div className="w-full h-4 bg-gray rounded-full overflow-hidden relative">
-      <div
-        className={`h-full bg-[linear-gradient(90deg,_#4557FF_0%,_#6E78FF_72.33%)] transition-all duration-1000 ease-in-out`}
-        style={{ width: getWidth() }}
-      />
-    </div>
-  );
-}
 
 export default function StrengthWeaknessFunnel() {
   const funnel = useFunnel(options);
